changed to float and added squared fn

+++ pages/api/subtract/minus.js
@@ -1,7 +1,7 @@
 export default function handler(req, res) {
   const { num1, num2 } = req.query
 
-  const result = parseInt(num1) - parseInt(num2)
+  const result = parseFloat(num1) - parseFloat(num2)
 
   res.status(200).json({ result: `${result}` })
 }
